fix(zig): cursor drift on j/k and improved font rendering quality
Two fixes:

1. Cursor drift: The Swift cursor overlay computed grid width from
   drawableSize / (cellWidth * contentsScale), which could differ from
   the actual grid width due to floating-point rounding. Each row
   accumulated +1 column of error. Fix: pass the actual grid_width from
   Zig to Swift and use it directly for cursor index calculation.

2. Font rendering: Switch from grayscale CGContext to RGBA. CoreText uses
   a superior LCD font rendering pipeline (subpixel smoothing) with RGBA
   contexts. Extract the green channel as alpha. Also enable subpixel
   positioning/quantization for sharper glyph placement, and store glyph
   bearing offsets as f64 instead of i32 to eliminate per-glyph rounding
   errors that caused vertical wobble between characters.

+++ zig/src/font/main.zig
@@ -19,8 +19,9 @@ pub const Glyph = struct {
     width: u32,
     height: u32,
     /// Bearing offsets for positioning relative to the baseline.
-    offset_x: i32,
-    offset_y: i32,
+    /// Bearing offsets in point space (fractional precision preserved).
+    offset_x: f64,
+    offset_y: f64,
 };
 
 /// Font face with glyph cache. Thread-safe for concurrent lookups
